Refactor: organize code into packages
Group classes into packages for better structure:
* command: Command classes
* storage: Storage and file I/O
* enums: shared enums
* exception: custom exceptions
* logic: Parser logic
* model: TaskList, Task, Deadline, Event, Todo
* ui: User display

Improves modularity, readability.
