+++ src/main/java/command/ResetCommand.java
@@ -1,8 +1,8 @@
 package command;
 
-import logic.Storage;
-import logic.TaskList;
-import logic.Ui;
+import storage.Storage;
+import model.TaskList;
+import ui.Ui;
 
 public class ResetCommand extends Command {
