+++ src/main/java/model/Event.java
@@ -1,4 +1,4 @@
-package task;
+package model;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
