+++ src/main/java/model/TaskList.java
@@ -1,6 +1,4 @@
-package logic;
-
-import task.Task;
+package model;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
