+++ text-ui-test/runtest.sh
@@ -20,7 +20,7 @@ then
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin Milo ../src/main/java/data/milo.txt < input.txt > ACTUAL.TXT
+java -classpath ../bin Milo ../src/main/java/storage/milo.txt < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
 cp EXPECTED.TXT EXPECTED-UNIX.TXT
